front-end-stream-lit/chatbotMV/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.25
langchain-core==0.3.74
langchain-openai==0.3.23
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langsmith==0.3.45
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
simple-salesforce==1.12.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
streamlit-chat==0.1.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.1
zeep==4.3.1
zstandard==0.23.0
